Remove Context and create "singleton" initialize-shutdown
The `libssh2_init` and `libssh2_exit` function should only be called once within an application. I do not think a "context" class is necessarily needed yet, and if the init and exit functions are called every time a context class is created, this prevents creating multiple SSH2 contexts in one application. Having more than one SSH2 context is rare, but I am not even sure if OOP is needed yet.

git-svn-id: file:///D/Users/cfield/Repos/LabSSH2/trunk@4 98f12691-ef58-ab42-8599-2451469670c7