Change from `libssh2` to `labssh2`
LabVIEW does not handle NULL in C interfaces very well. Some forum posts indicate passing a zero value for a sized or unsized pointer type works, but other forum posts indicate this is undefined behavior and could result in crashes. I was experiencing crashes. The universal truth among forum posts is to create a wrapper library around the shared library with the NULL value set in the wrapper library.

The libssh2 public API is very well designed, except for the macro functions that already include NULL values for the various functions. If these were real functions, all would be good but macro functions do NOT get included in the shared library. These are part of the include file (.h), which is ignored by LabVIEW. So, I forked the libssh2 project and started changing the macro functions to real functions. I call this fork, `labssh2`. This avoids having to create wrapper functions for all of the functions in the libssh2 library and basically reuse 99% of the project.

I named the "fork" to `labssh2` to avoid problems if users have libssh2 already installed. This ensures that the appropriate fork of libssh2 with the macro functions as real functions is used with LabVIEW.

git-svn-id: file:///D/Users/cfield/Repos/LabSSH2/trunk@7 98f12691-ef58-ab42-8599-2451469670c7