Add error handling
A custom error code file (LabSSH2-errors.txt) has been created and placed in the `C:\Program Files (x86)\National Instruments\LabVIEW 2017\user.lib\errors` folder along with a shortcut to the 64-bit install location (`C:\Program Files\National Instruments\LabVIEW 2017\user.lib\errors`. This is hopefully easier to manage than using the Error Code to Cluster VI, where I inevitably lose track of the errors and codes. I am not sure about the distribution of the file with a VIP or the VIPM. I will have to test it. The file is in an XML format, so any text editor can be used to modify it and there is also a built-in editor dialog to LabVIEW, but it is kind of cumbersome.

git-svn-id: file:///D/Users/cfield/Repos/LabSSH2/trunk@14 98f12691-ef58-ab42-8599-2451469670c7