Add host key API, which fixes the Known Hosts Check test, sort of
I was using the Host Key Hash, which I call the fingerprint, instead of the actual host key for the Check. Also, the OpenSSH server for Windows 10 uses EDCSA keys instead of SSH-RSA keys, so the known_hosts file had to be changed to use the SSH-RSA key. I followed this issue from GitHub:

https://github.com/jakob/Postico/issues/307

git-svn-id: file:///D/Users/cfield/Repos/LabSSH2/trunk@65 98f12691-ef58-ab42-8599-2451469670c7