Change name of shared library
I did not like the hyphens in the shared library names for the C project. Looking at the naming of other shared libraries on UNIX-like systems (Linux, macOS, and RT), hyphens are not really used in shared library (SO) names. They are usually named as compound words. To follow this convention, and because `libssh2lv` makes more sense than `libssh2-lv`, the shared libraries are renamed.