Refactor writing a "block" of bytes into a VI for the SFTP File API
A "block" of bytes is an array of bytes, or chuck, typically from a file that is of a user-defined size. The size is usually selected and optimized for reading a local file on disk in blocks, or chucks. This is different from the SSH or SFTP packet sizes used for writing the bytes to a remote file. This VI is useful for writing a file to a remote server in parts for implementing aborts, progress monitoring, and reducing local memory usage.