Refactor Write Block implementation to avoid needing hard-coded packet size constant
It is possible to not need the 30000 bytes constant, which is from the internals of the libssh2 library. This value is not publicly exposed by the libssh2 library, and it is not intended to be used or needed by users of the libssh2 library. It is an internal value. So, the Write Block VI is refactored to eliminate the need to "know" this value. All tests pass with this refactoring.