Add higher level, helper VI to the SFTP File API for getting the remote file size
Getting the remote file size is a very common need. Before, an attribute object would need to be created from a SFTP File object followed by using the SFTP Attributes API to read the file size, then it was important to destroy the SFTP Attributes object. This is very low-level and tedious. It would be nice to just be able to "get the file size" directly from the SFTP File object.