Change VI name
Instead of "Write Block", it is changed to "Write All" for the SFTP File API. This is to be more consistent with the Channel API and a little more clear. This VI blocks until all of the byte array is written, while the "Write" VI writes a single packet of bytes to the remote SFTP server based on the internal libssh2 packet sizes.