Add Read All VI to the Channel API
This is used to implement the SCP like functionality of downloading a file from a remote server. This is a common enough thing to do that it can be added to the API and it is consistent with the `Write All` VI. The implementation is largely based on the blocking SCP example from the libssh2's examples website.