Refactor control and VI name to be consistent without API names
The "Flags" control for the SFTP File class is renamed to "Options" to be consistent with the "Options" controls in other SFTP-related classes. The "Flags" name is derived from the libssh2 C library API, as these are implemented as bit masks in the C library, which is not really a LabVIEW concept. The "Options" name is more LabVIEW-centric than a "flag".