Change occurrences of `libssh2` to `libssh2lv`
The layers are from "bottom to top":

1. OpenSSL
2. libssh2
3. libssh2lv
4. libssh2-labview

The VIs in the libssh2-labview toolkit are calling into the libssh2lv shared
library, which is a wrapper around the libssh2 library, which is statically
compiled into the libssh2lv library along with the OpenSSL library. So,
references to "shared library" in the LabVIEW VI documentation should be to the
libssh2lv shared library. The structs exist there (libssh2lv). Many of the
structs have analogs in the libssh2 library, but not all.