Change package name and version
The package name is changed from "LIBSSH2" to "LIBSSH2 for LabVIEW". There are packages on the NI Tools Network and VIPM Community that are similar "wrappers" or implementations of libraries in other programming languages that follow the "X for LabVIEW" naming convention. For example, "ASCIIdoc for LabVIEW" and "SQLite for LabVIEW". The package name is changed to follow this loose naming convention and also (hopefully) further indicate the package uses the libssh2 C library and is NOT a pure LabVIEW implementation of the libssh2 library nor the SSH2 protocol.

The version number is changed from v1.0.0.X to v0.1.0.X to match the release versions. The NI Tools Network has a "No Beta" policy, but this project was recently published to the VIPM Community, where there does not appear to such a policy, and appears to be a more appropriate destination/publication platform for open source projects.