Fix example
The "Execute Multiple Commands with a Single Channel with a Raspberry Pi" example is fixed to actually work. There were several key modifications: (1) Set a Session timeout to avoid indefinitely waiting/stalling on reading from the channel, (2) creating a non-interactive shell by _not_ requesting a pseudo-terminal (pty) with the Terminal VI before request a shell channel, and (3) appending the Line Feed, a.k.a. Newline, character, `\n`, to each of the command strings. The newline character is needed to signal to the remote shell to actually execute the command. Creating a non-interactive shell eliminates local echoing and displaying a prompt. The Session timeout defaults to zero (0) in the libssh2 library, so when the length of the responses and login message are not known in advance, the VI can hang and wait indefinitely for more bytes in the Read buffer of the channel.