Fix example
The "Read-Execute-Print-Loop with a Raspberry Pi" example is fixed to avoid stalling and blocking on reads when there is no data available. This also incorporates the fixes and knowledge from fixing the other examples with appending a newline character but uses non-blocking mode instead of a Session timeout to avoid stalling. A non-interactive shell is created, too, but just to mention this is not an efficient or good implementation because it will appear to be an interactive shell to the user. This example should _NOT_ be used as a terminal or interactive shell. It is mostly for debugging and demonstrating simultaneous, non-blocking reading and writing.