Add optional connection timeout to connecting a session
The `Timeout (60000 ms)` optional input is added to the `Connect.vi` for the Session class. This means the VI no longer follows the connector pane/interface required for a property definition. The "Connect" property definition is removed along with the "Disconnect" property definition because the Connect/Disconnect are a pair and it is weird to have one and not the other as a property.